Hopefully fix the graphvizWithHandle bug
Ignore-this: 5cc11514205b9afbb89d14b0f48fa886
Thanks to Nikolas Mayr for pointing out where the error was: I thought
it was with runGraphvizCommand, but he found it was in the various
graphToGraph-style functions.

Not going to declare this closed yet; will wait for a while and see.

darcs-hash:20090917085520-97e08-209bc355831a94033405b5e3e6a5b452278dc97a